+++ tests/test_predict_ram.py
@@ -1,5 +1,5 @@
 """Short test for vmem prediction."""
-# Copyright 2018-2023 TeNPy Developers, GNU GPLv3
+# Copyright 2023 TeNPy Developers, GNU GPLv3
 
 
 import tenpy.algorithms as algo
@@ -24,7 +24,7 @@ def test_bosonic_model_TEBD():
     num_entries *= 5 # physical leg
     # expected MPS vmem:
     MPS_vmem = (num_entries * 8) / 1024
-    assert abs(engine.estimate_RAM(mini=0) - MPS_vmem) < 1e-10, "TEBD RAM did not match expectation (expected: %d, gotten:%d)" % (MPS_vmem, engine.estimate_RAM(mini=0))
+    assert abs(engine.estimate_RAM(mini=0) - int(MPS_vmem)) < 1e-10, "TEBD RAM did not match expectation (expected: %d, gotten:%d)" % (MPS_vmem, engine.estimate_RAM(mini=0))
 
 
 def test_bosonic_model_DMRG():
@@ -46,5 +46,5 @@ def test_bosonic_model_DMRG():
     MPO_vmem = (4**2*5**2 * (L-2) + 2*4*5**2 * 2) * 8 / 1024
     # add lanczos:
     RAM_lanczos = (3 * 5**2 * (99**2 * 4) + 2 * 99**2 * 5**2) * 8 / 1024
-    assert engine.estimate_RAM(mini=0) == (MPS_vmem+MPS_env_vmem+MPO_vmem+RAM_lanczos)/8, "DMRG RAM did not match expectation (expected: %d, gotten:%d)" % ((MPS_vmem+MPS_env_vmem+MPO_vmem+RAM_lanczos)/8, engine.estimate_RAM(mini=0))
+    assert engine.estimate_RAM(mini=0) == int((MPS_vmem+MPS_env_vmem+MPO_vmem+RAM_lanczos)/8), "DMRG RAM did not match expectation (expected: %d, gotten:%d)" % ((MPS_vmem+MPS_env_vmem+MPO_vmem+RAM_lanczos)/8, engine.estimate_RAM(mini=0))
